Calculating the Total Weight
But... why are we keeping track of the repLogs? Because now, we can more easily calculate the total weight. Before, we passed the Helper object the  element so that it could find all the tr elements and read the weight from them. We can simplify this! Instead, pass it our Array: this.repLogs.

At the bottom of this file, change the constructor() for Helper to have a repLogs argument. Set that on a repLogs property.

Below in calculateTotalWeight(), instead of using the {dollar_sign}wrapper to find all the tr elements, just pass this.repLogs to the static function. Inside of that, update the argument to repLogs.

Previously, _calculateWeights() would loop over the {dollar_sign}elements and read the data-weight attribute on each. Now, loop over repLog of repLogs. Inside, set totalWeight += repLog.totalWeightLifted.

Okay! Try that out! The table still loads... and the total still prints!