Introducing Set
But, ES2015 added one more new object that's related to all of this: Set. It's a lot like Array: it holds items... but with one important difference.

Open up play.js and set foods to an array.

Let's add gelato to the array and tortas. Clear everything else out.

And ya know what? Gelato is so good, we should add it again. At the bottom, log foods.

When you run the script, there are no surprises: gelato, tortas, gelato.