Ok, so do you love ES2015 yet? Man, I sure do. So... let's go use it immediately on our site! Oh wait.... didn't I mention that its features aren't supported by all browsers? Whoops. Yea, the latest version of Chrome supports everything... but if some of our users have older browsers, does this mean we can't use ES2015?
npm and yarn: Package Manager Battle!

Nope! You can! Thanks to an amazing tool called Babel. But before we talk about it, we need to do a little bit of installation. In the Node.js world, the package manager is called npm, and because you already installed Node, you already have it!

But wait! There's a disturbance in the Node.js package manager force. Yes, there is another... package manager... called Yarn! It's sort of a competitor to npm, but they work almost identically. For example, in PHP, we have a composer.json file. In Node, we will have a package.json... and you can use either npm or yarn to read it.

In other words,... you can use npm or yarn: they basically do the same thing. You could even have some people on your team using npm, and others using yarn.

We're going to use Yarn... because it's a bit more sophisticated. But, that means you need to install it! Check Yarn's Docs(https://yarnpkg.com/lang/en/docs/install/) for your install details.